Renamded `notFount` to `not-fount`
