+++ src/components/index.js
@@ -1,12 +1,12 @@
 import { module } from 'angular';
-import NotFound from './notFound/notFound';
+import Layout from './layout/layout';
+import NotFound from './not-found/not-found';
 import Auth from './auth/auth';
-import Dashboard from './dashboard/dashboard';
 
 const Components = module('app.components', [
+  Layout,
   NotFound,
   Auth,
-  Dashboard,
 ]);
 
 export default Components.name;
